requirements.txt
# Fichier requirements.txt
"""
Flask==2.3.3
Flask-Login==0.6.2
Flask-SQLAlchemy==3.1.1
boto3==1.34.0
python-dotenv==1.0.0
gunicorn==21.2.0
